+++ drift/CHANGELOG.md
@@ -3,6 +3,8 @@
 - __Breaking__: Custom SQL types have been replaced with first-party types,
  implement `DriftSqlType` instead.
 - __Breaking__: Rename `Constant` to `Literal`.
+- __Breaking__: Rename `innerJoin()` to `Join.inner`, `leftOuterJoin` to
+  `Join.leftOuter` and `crossJoin` to `Join.cross`.
 - The trailing `()` for column builders in Dart is no longer necessary.
   Replace `integer().autoIncrement()()` with `integer().autoIncrement()`.
 - Add `@DriftColumnDeclarationBuilder` annotation, allowing custom types to